Introduce a `.dockerignore` file
`.dockerignore` file improves the docker build times. For example, when a change is performed to a file that is not relevant for the docker context (for example README.md file is changed) then when `docker build` is executed this won't require a new image to be built again. Instead the existing image from the docker cache will be used/picked. `docker build` will run the build only when there is a change to a file that is relevant for the image build.