Add `.dockerignore` and perform minor cleanups (#22)
* Introduce a `.dockerignore` file

`.dockerignore` file improves the docker build times. For example, when a change is performed to a file that is not relevant for the docker context (for example README.md file is changed) then when `docker build` is executed this won't require a new image to be built again. Instead the existing image from the docker cache will be used/picked. `docker build` will run the build only when there is a change to a file that is relevant for the image build.

* Add only the required `charts/images.yaml` file, not the whole `charts/dir`

Currently the registry-cache container image has the whole charts dir in the container file system. This is not required. The extension is only using the `charts/images.yaml` file there and it is not applying any chart under the `charts` directory.

* Remove the exclusion for the deprecated var/func usage

Similar to https://github.com/gardener/gardener-extension-provider-aws/pull/672. The linter will now fail when deprecated var/func is used.

* Remove the duplicate license file

Currently 2 license files exist - `LICENSE` and `LICENSE.md`

* Clean up more gardener/gardener specific exclusions from `.golangci.yaml`