Drop the healthcheck controller
After https://github.com/gardener/gardener/pull/7462 gardenlet performs a health check for all ManagedResources (without the ones with class `seed`) in the Shoot control plane. The registry-cache extension deploys only a single ManagedResource and needs to report health checks for it. We no longer need to have a health check controller to do this. We can rely that gardenlet will health check the ManagedResource deployed by the extension.