Deflake the e2e tests
In prow the e2e tests fail because the scheduler-state.json file is not created. Locally, I am not able to reproduce the failure. But I checked the upstream distribution/distribution code and I see the following things that might cause the failures:
- On initial request an image seems to be pulled twice - once to return it to the requestor, and one more time to store it locally in the cache. The image pull for the local cache seems to happen async, in a dedicated goroutine. See https://github.com/distribution/distribution/blob/152af63ec5c569f074e9cf5d0e409d6928e034d8/registry/proxy/proxyblobstore.go#L111-L164
- Writes to the scheduler-state.json file are performed on given interval - 5 seconds currently. See https://github.com/distribution/distribution/blob/152af63ec5c569f074e9cf5d0e409d6928e034d8/registry/proxy/scheduler/scheduler.go#L21

This commit wraps the scheduler-state.json file handling in an `Eventually` to allow the test to wait until the above async operations are completed.