Change the `garbageCollectionEnabled` field to `garbageCollection.enabled`
The API change is done to make it easier in future to add new fields related to the garbage collection. For example there is a TTL field that we would like to expose when there is a distribution/distribution release with the PR that introduced the TTL future (ref https://github.com/distribution/distribution/pull/3880).