Make the cache size field immutable (#42)
* Make the cache size field immutable

Currently we are using StatefulSet volumeClaimTemplate to provision the PVC. However the corresponding StatefulSet field that controls the PVC size is immutable. It is not possible to resize/enlarge the underlying PVC through the StatefulSet spec. See https://github.com/kubernetes/enhancements/issues/661
This commit makes the cache size field immutable until we decide how to implement the resize.

* Address review comments