Explicitly set the PVC's `storageClassName` to `default` (#51)
* Explicitly set the PVC's `storageClassName` to `default`

In the registry-cache extension we deploy PVCs to the Shoot. We see cases where the PVC and StorageClass creation can be racy. In K8s < 1.26, when the default StorageClass is created after the PVC, the PVC hangs in Pending state with reason `no persistent volumes available for this claim and no storage class is set`. In K8s, there is a feature gate called `RetroactiveDefaultStorageClass` which makes possible the StorageClass from PVC to be set when the PVC is created before the StorageClass. See https://kubernetes.io/docs/concepts/storage/persistent-volumes/#retroactive-default-storageclass-assignment
In the registry-cache extension we decided to hard-code the StorageClass name to `default` for time being. When it is ensured that `RetroactiveDefaultStorageClass` is enabled by default in the supported K8s versions, then we can drop this.

* Vendor gardener/gardener@v1.80.4

gardener/gardener@v1.80.4 contains cherry-pick of https://github.com/gardener/gardener/pull/8565.

* Fix typo

* Update comment