Address TODO related to admission Pod labels
Right now the admission does not have a client and does not do requests against the virtual kube-apiserver or the runtime kube-apiserver.
In future we might need to introduce such client (when validating the registry cache credentials) but currently the admission components don't need access to DNS, virtual kube-apiserver or runtime kube-apiserver.